+++ src/Eternity/Constraints.cs
@@ -31,7 +31,7 @@ namespace Eternity
 			};
 		}
 
-		public static Constraints? Generate(IReadOnlyList<ImmutableArray<int>> pieceSides)
+		public static Constraints? Generate(IReadOnlyList<ImmutableArray<ulong>> pieceSides)
 		{
 			var sqrt = (int)Math.Round(Math.Sqrt(pieceSides.Count));
 			var dimensions = new Dimensions(sqrt, sqrt);
